+++ src/layout/section/cardsSection/CardsSection.unit.tsx
@@ -5,7 +5,7 @@ import { TripCard } from '../../../tripCard'
 import { CardsSection } from './CardsSection'
 
 const tripCardConfig = {
-  ariaLabel: 'Pick-up point: Paris, Drop-off point: Bordeaux, Departure time: 09:00, warning',
+  'aria-label': 'Pick-up point: Paris, Drop-off point: Bordeaux, Departure time: 09:00, warning',
   href: '/',
   itinerary: [
     {