+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
-[...]
+- **[BREAKING CHANGE]** Removed all `ariaX` props from components and switched to support all common `aria-x` attributes with `A11yProps` interface.
+  [...]
 
 # v37.2.0 (23/07/2020)